fix the code too dry -- remove the product information from the post product_routes -- add seed file to store product information
