+++ config/seed.js
@@ -1,0 +1,25 @@
+const products = [
+  {
+    imagePath: 'https://i.imgur.com/lnR1YSI.jpg',
+    title: 'REDESS Beanie Hat for Men and Women Winter Warm Hats Knit Slouchy Thick Skull Cap',
+    description: 'COMFORTABLE TO WEAR: This winter warm hat for men is made of high quality materials and well designed to fit men. With adjustable rubber band built inside perfect adjustability is available.',
+    price: 19.99
+  }, {
+    imagePath: 'https://i.imgur.com/WoyMZUw.jpg',
+    title: 'Lisianthus Women Belt Buckle Fedora Hat',
+    description: 'Breathable, lightweight and comfortable for all-day wear',
+    price: 16.45
+  }, {
+    imagePath: 'https://i.imgur.com/jrsvwuZ.jpg',
+    title: 'KBETHOS Lit Patch Finesse Self Made Dad Hat Baseball Cap Polo Style Unconstructed Cotton Adjustable',
+    description: 'ADJUSTABLE BUCKLE CLOSURE: One Size Fits Most, Comfortable Everyday Fit. Use the convenient metal closure to custom fit the baseball cap to your head and ensure maximum comfort and a secure fit at all times. The unisex design makes it suitably perfect for both men and women.',
+    price: 13.99
+  }, {
+    imagePath: 'https://i.imgur.com/GZYVYsi.jpg',
+    title: 'Prevue Pet Products Stainless Steel Playtop Bird Cage',
+    description: 'ADVANCED TECHNOLOGY: Columbia Unisex Bora Bora Booney fishing hat features our signature wicking fabric with sweat-wicking capabilities and UPF 50 sun protection technology that helps shield you from harmful UVA and UVB rays.',
+    price: 24.45
+  }
+]
+
+module.exports = products
